SP: Change ofs from int to size_t in g_save

+++ SP/code/game/g_save.c
@@ -78,7 +78,7 @@ typedef enum {
 } saveFieldtype_t;
 
 typedef struct {
-	int ofs;
+	size_t ofs;
 	saveFieldtype_t type;
 } saveField_t;
 
@@ -163,7 +163,7 @@ static saveField_t castStateFields[] = {
 //.......................................................................................
 // this is where we define fields or sections of structures that we should totally ignore
 typedef struct {
-	int ofs;
+	size_t ofs;
 	int len;
 } ignoreField_t;
 
@@ -204,7 +204,7 @@ static ignoreField_t castStateIgnoreFields[] = {
 // persistant data is optionally carried across level changes
 // !! WARNING: cannot save pointer or string variables
 typedef struct {
-	int ofs;
+	size_t ofs;
 	int len;
 } persField_t;
